+++ tests/tools/mypy/test_mypy_tool.py
@@ -1,6 +1,11 @@
+from pathlib import Path
 from unittest import SkipTest, TestCase
+from unittest.mock import patch
 
+from prospector.config import ProspectorConfig
+from prospector.finder import find_python
 from prospector.message import Location, Message
+from prospector.tools.exceptions import BadToolConfig
 
 try:
     from prospector.tools.mypy import format_message
@@ -9,6 +14,22 @@ except ImportError:
 
 
 class TestMypyTool(TestCase):
+    @staticmethod
+    def _get_config(profile_name: str) -> ProspectorConfig:
+        profile_path = Path(__file__).parent / f"test_profiles/{profile_name}.yaml"
+        with patch("sys.argv", ["prospector", "--profile", str(profile_path.absolute())]):
+            return ProspectorConfig()
+
+    def test_unrecognised_options(self):
+        finder = find_python([], [], True, Path(__file__).parent.absolute())
+        self.assertRaises(BadToolConfig, self._get_config("mypy_bad_options").get_tools, finder)
+
+    def test_good_options(self):
+        finder = find_python([], [], True, Path(__file__).parent.absolute())
+        self._get_config("mypy_good_options").get_tools(finder)
+
+
+class TestMypyMessageFormat(TestCase):
     def test_format_message_with_character(self):
         location = Location(path="file.py", module=None, function=None, line=17, character=2)
         expected = Message(source="mypy", code="error", location=location, message="Important error")