+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Just say no to bugs.
 
 * Stopped the ProfileValidator tool raising errors about ``pep8`` and ``pep257`` sections being unknown. Instead, they raise deprecated warnings.
 * Blending works again - for example, pylint and pycodestyle errors representing the same thing are combined. After renaming pep8 to pycodestyle, this only worked when using legacy names.
+* Unrecognised Mypy options now raise an exception instead of silently carrying on - `#455 <https://github.com/PyCQA/prospector/issues/455>`_
 
 **Tidyup**: